Add Config Alias Mechanism and Define New Aliases
In the HDFS configuration the following are aliases of each other:

    fs.default.uri   <deprecated>
    fs.defaultFS     <preferred>
    fs.default.name  <also deprecated I think>

In order to get all of these to set the same internal variable
for all of these, but the way the config works, assigning the
same variable to all of them in the code meant that the configured
value got overwritten by the default value most of the time.  The
solution was to declare the two less preferred values as aliases
of the preferred setting expicitly by associating a bool with the
definition that says whether the definition is original (ORIG) or
and alias (ALIAS).  Then, when reading in the config, if no
config value is found and the variable is not an alias, the
default is used, but if the variable is an alias the default is
not assigned.

Also, fixed a problem with cmake that was breaking my compile.
This worked on my (ancient) version of cmake that I have on my
actual build machine, but broke under docker when I was testing.